build: remove all typescript code
