--- .eslintrc.json
@@ -1,45 +1,0 @@
-{
-    "env": {
-        "node": true,
-        "commonjs": true,
-        "es2021": true
-    },
-    "parser": "@typescript-eslint/parser",
-    "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
-    "parserOptions": {
-        "ecmaVersion": "latest"
-    },
-    "plugins": ["eslint-plugin-import", "@typescript-eslint"],
-    "rules": {
-        "indent": [
-            "error",
-            4
-        ],
-        "linebreak-style": [
-            "error",
-            "unix"
-        ],
-        "quotes": [
-            "error",
-            "single"
-        ],
-        "semi": [
-            "error",
-            "always"
-        ],
-        "import/order": [
-            1,
-            {
-                "groups": ["external", "builtin", "internal", "sibling", "parent", "index"]
-            }
-        ],
-        "@typescript-eslint/no-unused-vars": ["error",
-            {
-                "vars": "all",
-                "args": "none",
-                "ignoreRestSiblings": false
-            }
-        ]        ,
-        "@typescript-eslint/no-explicit-any": ["off"]
-    }
-}
