+++ .gitignore
@@ -7,4 +7,4 @@ config.json
 config/*.json
 accessToken.json
 scripts/
-killbot-rust/target
+target/
