--- CHANGELOG.md
@@ -1,28 +1,0 @@
-2022-09-25
----
-
-- make system/region/constellation unsubscribeable
-- add new functionality to subscribe to a ship group
-- __attention: new commands will roll out within 24 hours__
- 
-2022-06-30
----
-
-- make alliance/corp/char killmails filterable by region/constellation/system (#5)
-- improved killmail embedings by fetching the open graph data from zkillboard url (#2)
-- colors for kills (green) / losses (red) on the embeds (#2)
-- __attention: new commands will roll out within 24 hours__
-
-2022-06-23
----
-
-- fix typo in corporation
-- add possibility to subscripe to region, constellation of system kills
-- add command to unsubscribe whole channel
-- __attention: new commands will roll out within 24 hours__
-
-2022-06-17
----
-
-- prevent message from sending twice to the same channel due to multiple subscriptions
-- try message owner of discord on channel permission problem when sending a killmail
