--- package.json
@@ -1,73 +1,0 @@
-{
-  "name": "s-connect-metrics-exporter",
-  "version": "0.1.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@discordjs/builders": "^0.13.0",
-    "@discordjs/rest": "^0.4.1",
-    "@types/open-graph-scraper": "^4.8.1",
-    "async-lock": "^1.3.1",
-    "axios": "^0.27.2",
-    "bin-pack": "^1.0.2",
-    "discord-api-types": "^0.33.2",
-    "discord.js": "^13.7.0",
-    "memory-cache": "^0.2.0",
-    "open-graph-scraper": "^4.11.1",
-    "prompt-sync": "^4.2.0",
-    "simple-oauth2": "^5.0.0",
-    "ts-node": "^10.9.2",
-    "ws": "^8.7.0"
-  },
-  "devDependencies": {
-    "@jest/expect": "^28.1.0",
-    "@types/async-lock": "^1.1.5",
-    "@types/bin-pack": "^1.0.3",
-    "@types/jest": "^27.0.2",
-    "@types/memory-cache": "^0.2.2",
-    "@types/prompt-sync": "^4.2.3",
-    "@types/simple-oauth2": "^5.0.7",
-    "@typescript-eslint/eslint-plugin": "^5.27.0",
-    "@typescript-eslint/parser": "^5.27.0",
-    "eslint": "^7",
-    "eslint-plugin-import": "^2.26.0",
-    "jest": "^27.3.1",
-    "nodemon": "^2.0.12",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.4.3"
-  },
-  "package": {
-    "install": "package.json"
-  },
-  "scripts": {
-    "start": "cd dist && node index",
-    "build": "./node_modules/typescript/bin/tsc",
-    "start.dev": "yarn && nodemon --ignore ./dist -e ts,twig,html,jpg,png,gif,svg,json --exec \"(yarn eslint . && ./node_modules/typescript/bin/tsc && cd dist && node index) || exit 1\"",
-    "test": "yarn jest",
-    "coverage": "yarn jest --coverage",
-    "lint": "yarn eslint 'src/**/*.ts'",
-    "lint:fix": "yarn eslint --fix 'src/**/*.ts'"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.ts?$": "ts-jest"
-    },
-    "testEnvironment": "node",
-    "testRegex": "/tests/.*\\.(test|spec)?\\.(ts|tsx|js)$",
-    "testPathIgnorePatterns": [
-      "<rootDir>/dist"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "moduleDirectories": [
-      "<rootDir>/node_modules",
-      "<rootDir>/src"
-    ]
-  }
-}
