+++ src/commands/diag.rs
