--- src/commands/helpCommand.ts
@@ -1,48 +1,0 @@
-import {SlashCommandBuilder} from '@discordjs/builders';
-import {CommandInteraction} from 'discord.js';
-import * as util from 'util';
-import {AbstractCommand} from './abstractCommand';
-import {ZKillSubscriber} from '../zKillSubscriber';
-
-export class HelpCommand extends AbstractCommand {
-    protected name = 'zk-activity-diag';
-
-    async executeCommand(interaction: CommandInteraction): Promise<void> {
-        const sub = ZKillSubscriber.getInstance();
-        if (!interaction.inGuild()) {
-            console.log(util.inspect(interaction, { showHidden: false, depth: 5 }));
-            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-            // @ts-ignore
-            await interaction.reply('Diagnostics is not possible in PM!');
-            return;
-        }
-        const subs = await sub.listGuildChannelSubscriptions(interaction.guildId, interaction.channelId);
-
-        const subs_str = util.inspect(subs, { showHidden: false, depth: 5 } );
-        const log = [
-            'List of subscriptions for guild',
-            interaction.guildId,
-            'channel',
-            interaction.channelId,
-            'is',
-            subs_str
-        ].join(' ');
-
-        console.log(log);
-
-        await interaction.reply({
-            content: log,
-            ephemeral: true
-        });
-    }
-
-    getCommand(): SlashCommandBuilder {
-        return new SlashCommandBuilder().setName(this.name)
-            .setDescription('Help and diagnostics');
-
-    }
-
-}
-
-
-
