+++ src/commands/subscribe.rs
